Refactor: Actualización del tipo de dato del precio a Double en toda la app y adición de formato de moneda colombiana.

+++ app/src/main/java/com/univalle/inventoryapp/viewmodel/InventoryViewModel.kt
@@ -69,9 +69,9 @@ class InventoryViewModel(application: Application) : AndroidViewModel(applicatio
         }
     }
 
-    fun totalProducto(price: Int, quantity: Int): Double {
+    fun totalProducto(price: Double, quantity: Int): Double {
         val total = price * quantity
-        return total.toDouble()
+        return total
     }
 
 }
